pyproject.toml
[project]
name = "ferry"
version = "0.1.0"
requires-python = ">=3.14"

[tool.uv.workspace]
members = ["utils", "backend", "action"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "moto[dynamodb,apigateway,stepfunctions]>=5.0",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "boto3-stubs[dynamodb]>=1.34",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["utils/src", "backend/src", "action/src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"tests/**/test_*.py" = ["TCH002"]
"scripts/*.py" = ["E402"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
